requirements.txt
pyyaml
requests
scapy>=2.5.0; platform_system=="Linux"
